Improve client-entry.ts testability
This extracts the initialization logic for Termbox's config and store
services which are difficult and not that useful to test into a separate
`initializeConfigAndDefaultServices` function. The function in
client-entry.ts is also changed from a `void` type to a `Promise<void>`
so that tests can "wait" for it to complete.

Extracting this default initialization logic is done in preparation for
configurable services to be injected into the client-entry function.

Bug: T342234
Change-Id: Ieb29c9d5f8d9559a2606dc65be6215039055d834