Export relevant interfaces alongside init function type
This adds a new `client-entry-type.ts` file in order to avoid exporting
interfaces from `client-entry.ts` directly, which a) causes warnings at
build time (see https://github.com/webpack/webpack/issues/7378) and also
makes the exported interfaces appear in the js dist file, which we don't
want. The only purpose of this file is to export the type of the init
function from `client-entry.ts` and other interfaces that may be useful
for consumers.

Bug: T342235
Change-Id: I71fb071bcd7dafa817e22ff04d8bfcefcd9a26e1