Replace WikimediaUI Base with Codex design tokens
This patch replaces the WikimediaUI Base library with the new
Codex design tokens library.
Codex tokens are mostly a drop-in replacement for the old one.
But there are deprecated variable names (token keys) that need to be
adapted to the new ones.
It's also removing that extra leg of replacing Less source vars with
SASS variables as Codex provides them out of box. Removing build
dependencies around that.

For code hygiene reasons, this patch also removes a number of local
color variables as exemplary new usage for tokens.
Only subtle visual change is replacing `$color-moderate-red: #a55858;`
with `$color-error: #d73333` as the agreed-upon color.
`$color-light-gray` was unused and removed.
The message colors have changed on borders for example too.
All other replacements are just name changes for a unified nomenclature.

Replacing also a number of static values mainly on properties, where
most other occurrences had already majority of variable usage. See
`border-radius` for example.

Bug: T331403
Bug: T332120
Change-Id: Ice1170b552dc12df5e95646bd218d166144712ce