Redirect from termbox if edit creates new temporary user
In preparation for IP Masking, we need to process redirects that
we receive from the API (T357024) in the Mobile Termbox after we
edit a term when logged out and temp-user is enabled. Depends on
changes to the RepoApiWritingEntityRepository in Wikibase itself,
but this submodule update will be applied when the dependent
changes get introduced.

Bug: T357151
Change-Id: I1b8389dc5cd5557e209a3c68b4fc3d38888466d1