Update anonymous edit warning for temp user case
Now that we are introducing temporary user support, it is no longer
the case that making an anonymous edit will expose the users' IP
addresses. As we have done with other components, we adjust the
warning message here to depend on the state of the repo where the
Termbox is running.

Bug: T357404
Change-Id: Id0665816e57e2dfaa6d669ce68489f0fc1e9e1b5