Remove disabled lib version check for Vue
This removes a disabled CI test which was supposed to ensure that the
MediaWiki Vue version matches the one that termbox uses. The test did
not survive the Vue3 upgrade and there is no easy fix to re-enable it.

We've decided to remove the disabled test, because client side rendering
and server side rendering are operating independently. In the future we
will rely on notifications from the MediaWiki team regarding major
version updates of the Vue library.

Bug: T357269
Change-Id: I2fe5ac72aece3fcfa8ffe05cde8fdfba27dbeb5f