Prepare for wdio-mediawiki v5
Change all the files and imports to ESM (via a new package.json in
tests/selenium/); because legacy ESLint can’t load ESM .eslintrc.js
files, rename that file to .cjs.

Note that the wdio-wikibase peer dependency in package.json is at this
point outdated to the point of being almost meaningless; with the ESM
imports, we certainly need wdio-wikibase ^7.0.0. However, I have been
unable to get CI to work with a bumped peer dependency (see change
I79d740a418 for an attempt), so leave it outdated for now.

Bug: T406844
Change-Id: Iffa95bfaaa6ee5723fbdd52a5a6ab586c19c8efd